requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pygments>=2.15.0
aiohttp>=3.8.0
asyncio>=3.4.3 